Remove hard-coding in student method for application controller Display homework for student submissions Update menu options to go somewhere or be disabled

+++ db/migrate/20200707042546_change_homework_submission_user.rb
@@ -1,6 +1,6 @@
 class ChangeHomeworkSubmissionUser < ActiveRecord::Migration[6.0]
   def change
-    remove_column :homework_submissions, :course_session_user_id
+    remove_column :homework_submissions, :course_session_user_id if column_exists? :homework_submissions, :course_session_user_id
     add_column :homework_submissions, :course_session_participant_id, :bigint
   end
 end
